requirements.txt
# Intentionally minimal for v1. Add dependencies only when needed.
pdf2image
shapely
rtree
Pillow
google-generativeai
python-pptx
